vega/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
nuscenes-devkit>=1.1.11
opencv-python>=4.8.0
numpy>=1.24.0
albumentations>=1.3.0
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
matplotlib>=3.7.0
tqdm>=4.65.0
